// src/reports/constants/report-types.ts
export enum ReportsTypes {
    CERTIFICATIONS_USERS = 'Certifications - Users',
    COURSES_USERS = 'Courses - Users',
    ECOMMERCE_TRANSACTION = 'Ecommerce - Transactions',
    GROUPS_COURSES = 'Groups/Branches - Courses',
    SURVEYS_INDIVIDUAL_ANSWERS = 'Surveys - Individual Answers',
    USERS = 'Users',
    USERS_BADGES = 'Users - Badges',
    USERS_CERTIFICATIONS = 'Users - Certifications',
    USERS_CLASSROOM_SESSIONS = 'Users - Classroom Sessions', // Users - ILT Sessions
    USERS_COURSES = 'Users - Courses',
    USERS_ENROLLMENT_TIME = 'Users - Course Enrollment Time',
    USERS_EXTERNAL_TRAINING = 'Users - External Training',
    USERS_LEARNINGOBJECTS = 'Users - Learning Objects', // Users - Training Material
    USERS_LP = 'Users - Learning Plans',
    USERS_WEBINAR = 'Users - Webinar Sessions', // To be deprecated
    ASSETS_STATISTICS = 'Assets - Statistics',
    USER_CONTRIBUTIONS = 'User - Contributions', // User - Assets
    QUERY_BUILDER_DETAIL = 'Query - Builder',
    VIEWER_ASSET_DETAILS = 'Viewer - Asset Details',
    SESSIONS_USER_DETAIL = 'Sessions - Users Statistics',
    MANAGER_USERS_COURSES = '01',
    MANAGER_USERS_LP = '02',
    MANAGER_USERS_CERTIFICATIONS = '03',
    MANAGER_USERS_CLASSROOM_SESSIONS = '04',
    LP_USERS_STATISTICS = 'Learning plans - Users Statistics',
}